refactor(foundation): unify Protocol + incidents types and migrate http forecasting (A15)
Move two more families of types to `mockforge-foundation` so both
`mockforge-core` and `mockforge-contracts` share the exact same
definitions:

1. Protocol enum: previously duplicated at
   `mockforge_core::protocol_abstraction::Protocol` and
   `mockforge_contracts::protocol::Protocol`. Both now re-export from
   `mockforge_foundation::protocol::Protocol`.

2. Incident types: `IncidentStatus`, `IncidentType`, `IncidentSeverity`,
   `DriftIncident` (+ methods `new`, `acknowledge`, `resolve`, `close`,
   `link_external_ticket`), `ExternalTicket`, `IncidentQuery`.
   Previously duplicated byte-for-byte at
   `mockforge_core::incidents::types` and
   `mockforge_contracts::incidents::types`. Both now re-export from
   `mockforge_foundation::incidents_types`.

Also aligns contracts' `contract_drift::{fitness, consumer_mapping}` to
re-export the same data types (FitnessTestResult, AppType,
ConsumingApp, SDKMethod, ConsumerMapping, ConsumerImpact) from
foundation, so cross-crate method signatures match.

With Protocol + DriftIncident unified, http's forecasting handlers,
service, and router wiring can migrate from
`mockforge_core::contract_drift::forecasting::*` to
`mockforge_contracts::contract_drift::forecasting::*`.

Drops another ~43 deprecation warnings from http (511 → 468). All
foundation (42) + core (1818) + contracts + http (574) tests pass;
workspace clippy clean.
