+++ crates/mockforge-core/src/protocol_abstraction/mod.rs
@@ -27,51 +27,9 @@ pub use streaming::{
     StreamingProtocolRegistry,
 };
 
-/// Protocol type enumeration for multi-protocol support
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum Protocol {
-    /// HTTP/REST protocol for RESTful APIs
-    Http,
-    /// GraphQL protocol for GraphQL APIs
-    GraphQL,
-    /// gRPC protocol for gRPC services
-    Grpc,
-    /// WebSocket protocol for real-time bidirectional communication
-    WebSocket,
-    /// SMTP/Email protocol for email communication
-    Smtp,
-    /// MQTT protocol for IoT messaging and pub/sub
-    Mqtt,
-    /// FTP protocol for file transfer operations
-    Ftp,
-    /// Kafka protocol for distributed event streaming
-    Kafka,
-    /// RabbitMQ/AMQP protocol for message queuing
-    RabbitMq,
-    /// AMQP protocol for advanced message queuing scenarios
-    Amqp,
-    /// TCP protocol for raw TCP connections
-    Tcp,
-}
-
-impl fmt::Display for Protocol {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Protocol::Http => write!(f, "HTTP"),
-            Protocol::GraphQL => write!(f, "GraphQL"),
-            Protocol::Grpc => write!(f, "gRPC"),
-            Protocol::WebSocket => write!(f, "WebSocket"),
-            Protocol::Smtp => write!(f, "SMTP"),
-            Protocol::Mqtt => write!(f, "MQTT"),
-            Protocol::Ftp => write!(f, "FTP"),
-            Protocol::Kafka => write!(f, "Kafka"),
-            Protocol::RabbitMq => write!(f, "RabbitMQ"),
-            Protocol::Amqp => write!(f, "AMQP"),
-            Protocol::Tcp => write!(f, "TCP"),
-        }
-    }
-}
+// `Protocol` is re-exported from `mockforge_foundation::protocol` so both
+// core and `mockforge-contracts` share the canonical enum.
+pub use mockforge_foundation::protocol::Protocol;
 
 /// Message pattern enumeration for different communication patterns
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
