+++ crates/mockforge-foundation/src/lib.rs
@@ -14,8 +14,10 @@ pub mod contract_diff_types;
 pub mod contract_drift_types;
 pub mod encryption_error;
 pub mod error;
+pub mod incidents_types;
 pub mod intelligent_behavior;
 pub mod multi_tenant_types;
+pub mod protocol;
 pub mod state_machine;
 pub mod workspace_promotion;
