+++ crates/mockforge-http/src/lib.rs
@@ -2261,7 +2261,7 @@ pub async fn build_router_with_chains_and_multi_tenant(
         use crate::handlers::forecasting::{forecasting_router, ForecastingState};
         use crate::handlers::semantic_drift::{semantic_drift_router, SemanticDriftState};
         use crate::handlers::threat_modeling::{threat_modeling_router, ThreatModelingState};
-        use mockforge_core::contract_drift::forecasting::{Forecaster, ForecastingConfig};
+        use mockforge_contracts::contract_drift::forecasting::{Forecaster, ForecastingConfig};
         use mockforge_core::contract_drift::threat_modeling::{
             ThreatAnalyzer, ThreatModelingConfig,
         };
